backend_code/requirements.txt
Package             Version
------------------- -------
Django              2.2    
django-cors-headers 3.2.1  
djangorestframework 3.11.0 
Pillow              7.1.1  
pip                 10.0.1 
PyMySQL             0.9.3  
pytz                2019.3 
setuptools          39.1.0 
sqlparse            0.3.1  
